How to get da points holy moley

+++ PishiStirayNET/Infrastructure/Converters/IntToBrushConverter.cs
@@ -14,7 +14,7 @@ namespace PishiStiray.Infrastructure.Converters
     {
         public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
         {
-            float? number = (float?)value;
+            sbyte? number = (sbyte?)value;
 
             if (number == null || number <= 15)
             {
